requirements.txt
# SImS - Semantic Image Search
# Core Dependencies

# Web framework
fastapi>=0.124.0
uvicorn>=0.38.0
pydantic>=2.12.0

# HTTP client (for Ollama API calls)
httpx>=0.28.0

# Image processing
pillow>=12.0.0
pillow-heif>=1.1.0
rawpy>=0.25.0

# Metadata extraction
exifread>=3.5.0

# Vector database
chromadb>=1.3.0

# Utilities
python-dotenv>=1.2.0

# Testing
pytest>=9.0.0
pytest-cov>=7.0.0
pytest-asyncio>=1.3.0
